orders

+++ src/main/java/by/javatr/library/command/impl/DeleteBookCommand.java
@@ -4,11 +4,15 @@ import by.javatr.library.command.Command;
 import by.javatr.library.entity.Book;
 import by.javatr.library.exception.ServiceException;
 import by.javatr.library.service.BookService;
+import by.javatr.library.util.Constants;
+import org.apache.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
+import java.util.Optional;
 
 public class DeleteBookCommand implements Command {
+    private final static Logger LOGGER = Logger.getLogger(DeleteBookCommand.class);
     private BookService bookService;
 
     public DeleteBookCommand(BookService bookService) {
@@ -17,21 +21,25 @@ public class DeleteBookCommand implements Command {
 
     @Override
     public String execute(HttpServletRequest request) {
-        String page = "/WEB-INF/jsp/main.jsp";
+        String page = Constants.MAIN;
         String title = request.getParameter("title");
         try {
-            Book book = bookService.findByTitle(title);
-            int numberOfInstances = book.getNumberOfInstances();
-            if (numberOfInstances > 1) {
-                book.setNumberOfInstances(numberOfInstances - 1);
-                bookService.update(book);
-            } else {
-                bookService.delete(book);
+            Optional<Book> bookOptional = bookService.findByTitle(title);
+            if (bookOptional.isPresent()){
+                Book book = bookOptional.get();
+                int numberOfInstances = book.getNumberOfInstances();
+                if (numberOfInstances > 1) {
+                    book.setNumberOfInstances(numberOfInstances - 1);
+                    bookService.update(book);
+                } else {
+                    bookService.delete(book);
+                }
+                List<Book> allBooks = bookService.findAll();
+                request.setAttribute("books", allBooks);
             }
-            List<Book> allBooks = bookService.findAll();
-            request.setAttribute("books", allBooks);
         } catch (ServiceException e) {
-            page = "/WEB-INF/jsp/error.jsp";
+            LOGGER.error(e.getMessage(), e);
+            page = Constants.ERROR;
         }
         return page;
     }
